* print-rt6.c: make IPv6 routing header printing work with new 2292bis   API. * print-bgp.c: improve options printing.  ugly code exists for   unaligned option parsing (need some fix). * const poisoning in SMB decoder. * make dump format back to original.  someone may want to add an   option to do ascii printing, but keep the default behavior as is   for scripts used in many places. * -Wall -Werror clean checks.

+++ print-frag6.c
@@ -74,16 +74,16 @@ frag6_print(register const u_char *bp, register const u_char *bp2)
 	TCHECK(dp->ip6f_offlg);
 
 	if (vflag) {
-		printf("frag (0x%08x:%d|%d)",
-		       ntohl(dp->ip6f_ident),
+		printf("frag (0x%08x:%d|%ld)",
+		       (u_int32_t)ntohl(dp->ip6f_ident),
 		       ntohs(dp->ip6f_offlg & IP6F_OFF_MASK),
 		       sizeof(struct ip6_hdr) + ntohs(ip6->ip6_plen) -
-			       (bp - bp2) - sizeof(struct ip6_frag));
+			       (long)(bp - bp2) - sizeof(struct ip6_frag));
 	} else {
-		printf("frag (%d|%d)",
+		printf("frag (%d|%ld)",
 		       ntohs(dp->ip6f_offlg & IP6F_OFF_MASK),
 		       sizeof(struct ip6_hdr) + ntohs(ip6->ip6_plen) -
-			       (bp - bp2) - sizeof(struct ip6_frag));
+			       (long)(bp - bp2) - sizeof(struct ip6_frag));
 	}
 
 #if 0
